requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
dnspython==1.16.0
idna==2.9
lxml==4.5.0
mysql-connector-python==8.0.19
protobuf==3.6.1
pymongo==3.10.1
python-wordpress-xmlrpc==2.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
urllib3==1.25.8
